projects/element-secretary/requirements.txt
# Element Secretary Dependencies

# Matrix client
matrix-nio[e2e]>=0.24.0

# HTTP client for backend calls
httpx>=0.27.0

# Async support
aiofiles>=23.2.1

# Environment variables
python-dotenv>=1.0.0

# SQLite (built-in, but aiosqlite for async)
aiosqlite>=0.19.0
